requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pandas
numpy
scikit-learn
xgboost
prophet
matplotlib
seaborn
pytest
httpx
jupyter
faker
pydantic
